relocated UI tests and UI dto's into UI-packages
