+++ src/test/java/pages/AddMoneyPage.java
@@ -1,7 +1,7 @@
 package pages;
 
 import com.codeborne.selenide.Selenide;
-import dto.User;
+import dto.ui.User;
 import io.qameta.allure.Step;
 import lombok.extern.log4j.Log4j2;
 import org.openqa.selenium.By;
