+++ src/test/java/tests/ui/AddMoneyTest.java
@@ -1,6 +1,6 @@
-package tests;
+package tests.ui;
 
-import dto.User;
+import dto.ui.User;
 import io.qameta.allure.Description;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Link;
@@ -8,7 +8,7 @@ import io.qameta.allure.Owner;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-import static dto.UserFactory.getUser;
+import static dto.ui.UserFactory.getUser;
 import static org.testng.AssertJUnit.assertEquals;
 
 public class AddMoneyTest extends BaseTest {
